Assignment-2/myproject/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
ruff==0.6.9
sqlparse==0.5.1
